agents/risk_assessment/agent.py
# agents/risk_assessment/agent.py
#
# Risk Assessment Agent — evaluates location-specific wildfire risk.

from google.adk.agents import LlmAgent
from google.adk.tools import google_search

root_agent = LlmAgent(
    model="gemini-2.0-flash",
    name="wildfire_risk_agent",
    description="Evaluates location-specific wildfire risk using fire history, terrain, vegetation, and weather patterns.",
    instruction="""You are a wildfire risk assessment specialist.

Given a property location, evaluate its wildfire risk across these dimensions:

1. **Fire History**: Recent fires within 50 miles, historical burn frequency for the area.
2. **Terrain Risk**: Slope, aspect, elevation. Uphill properties face higher risk.
3. **Vegetation Zone**: Type and density of surrounding vegetation. WUI zone classification.
4. **Weather Patterns**: Seasonal fire weather (Santa Ana winds, dry lightning, drought conditions).
5. **Overall Risk Score**: Rate as Low / Moderate / High / Very High / Extreme with justification.

Use google_search to find:
- Recent wildfire history for the specific area
- NOAA weather and drought data
- State fire hazard severity zone maps

Be specific about the location. Reference actual fires, actual weather patterns, actual zone classifications.
Output a structured risk profile with all 5 dimensions.
""",
    tools=[google_search],
)
